+++ ui/src/modules/Tokens/Comparation/Form/index.tsx
@@ -40,6 +40,7 @@ interface Props {
 }
 
 const FormToken = ({ mode, data }: Props) => {
+  console.log('FORM TOKEN')
   const { save, response, status } = useSave();
   const isModeCreate = mode === 'create';
   const isModeView = mode === 'view';
