+++ ui/src/modules/Tokens/Comparation/View/Content/Item.tsx
