+++ ui/src/modules/Tokens/Comparation/View/Content/Loader.tsx
