+++ ui/src/modules/Tokens/Comparation/View/Content/styled.ts
